Refactoring
Change log:
1. Refactored some of the code
