+++ Space Invaders/classes/Game/Colors.py
@@ -1,0 +1,9 @@
+#Define the COLORS
+WHITE = (255,255,255)
+GREY = (60,60,60)
+BLACK = (0,0,0)
+RED = (255,0,0)
+BLUE = (0,0,255)
+GREEN = (0,128,0)
+LIME = (0,255,0)
+YELLOW = (255,255,0)
